+++ src/common/model/index.ts
@@ -2,7 +2,6 @@ export * from './expression';
 export * from './expression-type';
 export * from './failure';
 export * from './match';
-export * from './node-evaluation';
 export * from './operator-associativity';
 export * from './operator-type';
 export * from './ord';