+++ src/common/runtimes/index.ts
@@ -1,2 +1,0 @@
-export * from './binary-operator';
-export * from './function';