+++ src/dto/config/arg-demand.ts
@@ -1,0 +1,4 @@
+export enum ArgDemand {
+  Optional = 'Optional',
+  Required = 'Required',
+}