+++ src/dto/config/function-arg.ts
@@ -1,0 +1,8 @@
+import { ValueType } from '../../common/model';
+import { ArgDemand } from './arg-demand';
+
+export interface IFunctionArg {
+  readonly demand: ArgDemand;
+  readonly label: string;
+  readonly type: ValueType;
+}