+++ src/testers/operators/gt.ts
@@ -1,4 +1,0 @@
-import { Ordering } from '../../common/model';
-import { relational } from './relation';
-
-export const gt = relational(ord => ord === Ordering.Gt);